+++ SampleApps/PingExample/PingExample/Collectors/LabelView.swift
@@ -1,8 +1,8 @@
-// 
+//
 //  LabelView.swift
 //  PingExample
 //
-//  Copyright (c) 2025 Ping Identity Corporation. All rights reserved.
+//  Copyright (c) 2025 - 2026 Ping Identity Corporation. All rights reserved.
 //
 //  This software may be modified and distributed under the terms
 //  of the MIT license. See the LICENSE file for details.
@@ -14,12 +14,23 @@ import PingDavinci
 
 struct LabelView: View {
     var field: LabelCollector
-    
+
     var body: some View {
         HStack {
-            Text(field.content)
+            labelContent
                 .font(.subheadline)
                 .frame(maxWidth: .infinity, alignment: .center)
+        }
+    }
+
+    // MARK: - Label Content
+
+    @ViewBuilder
+    private var labelContent: some View {
+        if let richContent = field.richContent {
+            Text(RichTextBuilder.build(from: richContent))
+        } else {
+            Text(field.content)
         }
     }
 }